refactor: standardize file naming conventions
