+++ frontend/src/components/forms/ui/StudentFormUI.tsx
@@ -1,4 +1,4 @@
-import CommonButton from "@/components/common_button/CommonButton";
+import CommonButton from "@/components/common-button/CommonButton";
 import { useAppNavigation } from "@/utils/navigator";
 import { Check } from "lucide-react";
